Use last_seq not update_seq
I had assumed that the update_seq value from GET /dbname would always
match the last_seq value in the changes feed. It turns out that some
changes, like PUT /dbname/_revs_limit and PUT /dbname/_security can
increase update_seq without making an entry in the changes feed.

The upshot of this discrepancy is that c-l will hang waiting for an
update that will never arrive.

The fix is to read the last_seq value from the changes feed instead
and then wait for the indexer to catch up to that value.